+++ tests/TestCase/Command/IndexCommandTest.php
@@ -42,7 +42,8 @@ class IndexCommandTest extends TestCase
     {
         $files = $this->createSomeBackups();
 
-        $this->exec($this->command);
+        $this->deprecated(fn() => $this->exec($this->command));
+
         $this->assertExitSuccess();
         $this->assertOutputContains('Backup files found: 3');
 
@@ -88,18 +89,13 @@ class IndexCommandTest extends TestCase
         $this->assertSame(array_map(callback: 'basename', array: $files), $matches[1]);
     }
 
-    /**
-     * Test for `execute()` method, with `--reverse` option.
-     *
-     * @test
-     * @uses \DatabaseBackup\Command\IndexCommand::execute()
-     */
     #[Test]
     public function testExecuteWithReverseOption(): void
     {
         $files = array_reverse($this->createSomeBackups());
 
-        $this->exec($this->command . ' --reverse');
+        $this->deprecated(fn() => $this->exec($this->command . ' --reverse'));
+
         $this->assertExitSuccess();
         $this->assertOutputContains('Backup files found: 3');
 
@@ -124,15 +120,11 @@ class IndexCommandTest extends TestCase
         $this->assertSame(array_map(callback: 'basename', array: $files), $matches[1]);
     }
 
-    /**
-     * Test for `execute()` method, with no backup files.
-     *
-     * @uses \DatabaseBackup\Command\IndexCommand::execute()
-     */
     #[Test]
     public function testExecuteWithNoFiles(): void
     {
-        $this->exec($this->command);
+        $this->deprecated(fn() => $this->exec($this->command));
+
         $this->assertExitSuccess();
         $this->assertOutputContains('Backup files found: 0');
         $this->assertErrorEmpty();